+++ Gemfile
@@ -1,0 +1,5 @@
+# A sample Gemfile
+source "https://rubygems.org"
+
+gem 'sqlite3'
+gem 'artii'
