+++ db_connection.rb
@@ -1,0 +1,3 @@
+# Testで使うこともあるので
+def App.dbm;  DatabaseManagement.new end
+def App.db;   SQLite3::Database.new("speech.db") end
